+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Datasplit of test, train and val, is not done sepparately for ionmodes anymore.
+
+### Changed
+- Settings include file name of spectra now. This makes tracking of runs more easily and more flexibility for results folder. 
 
 ## [2.6.0] - 2025-12-05
 ### Changed